+++ src/Sher/Wap/Action/Shop.php
@@ -527,7 +527,7 @@ class Sher_Wap_Action_Shop extends Sher_Wap_Action_Base {
           $is_snatched = true;
           $snatch_product_id = $product_data['_id'];
           // 如果抢购价为0,设置订单状态为备货
-          if((int)$pay_money==0){
+          if((float)$pay_money==0){
             $order_info['status'] = Sher_Core_Util_Constant::ORDER_READY_GOODS;
             $order_info['is_payed'] = 1;              
           }
@@ -570,7 +570,7 @@ class Sher_Wap_Action_Shop extends Sher_Wap_Action_Base {
     }
 		
     //如果是抢购并且为0元抢，无需支付，跳到我的订单页
-    if($is_snatched && (int)$pay_money==0){
+    if($is_snatched && (float)$pay_money==0){
       $next_url = Doggy_Config::$vars['app.url.wap'].'/my/order_view?rid='.$rid;
     }else{
       $next_url = Doggy_Config::$vars['app.url.wap'].'/shop/success?rid='.$rid;
